Move reset combination
Don't make reset where enter is on layer 2.

My typical usage pattern is to press enter after using the arrow
keys. This would require a layer switch, but if I forget, I end up
triggering the reset and need to wait 8 seconds for the keyboard to
return.

This moves the reset key left by one and puts the enter key back in
layer 2.